requirements.txt
asgiref==3.10.0
autopep8==2.0.4
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
dj-database-url==0.5.0
Django==4.2
filelock==3.19.1
fsspec==2025.10.0
googlesearch-python==1.2.4
gunicorn==20.1.0
h11==0.16.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.2
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pillow==10.4.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
python-dateutil==2.9.0.post0
python-decouple==3.4
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
reportlab==4.4.4
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
six==1.17.0
soupsieve==2.8
sqlparse==0.4.1
stripe==8.8.0
sympy==1.14.0
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.8.0
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
whitenoise==6.8.1
